linux_pwm_out fix sign-compare

+++ src/drivers/linux_pwm_out/linux_pwm_out.cpp
@@ -285,7 +285,7 @@ void task_main(int argc, char *argv[])
 			_controls[0].control[2] = 0.f;
 			int channel = rc_channels.function[rc_channels_s::RC_CHANNELS_FUNCTION_THROTTLE];
 
-			if (ret == 0 && channel >= 0 && channel < sizeof(rc_channels.channels) / sizeof(rc_channels.channels[0])) {
+			if (ret == 0 && channel >= 0 && channel < (int)(sizeof(rc_channels.channels) / sizeof(rc_channels.channels[0]))) {
 				_controls[0].control[3] = rc_channels.channels[channel];
 
 			} else {
@@ -347,7 +347,7 @@ void task_main(int argc, char *argv[])
 					pwm_value = _pwm_min;
 				}
 
-				for (int i = 0; i < _outputs.noutputs; ++i) {
+				for (uint32_t i = 0; i < _outputs.noutputs; ++i) {
 					pwm[i] = pwm_value;
 				}
